Move builder back to index.ts
